+++ src/main/java/com/ghostchu/peerbanhelper/module/impl/webapi/PBHChartController.java
@@ -24,8 +24,6 @@ import java.net.UnknownHostException;
 import java.sql.Timestamp;
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ExecutorService;
@@ -84,41 +82,8 @@ public class PBHChartController extends AbstractFeatureModule {
 
     private void handleTrafficClassic(Context ctx) throws Exception {
         var timeQueryModel = WebUtil.parseTimeQueryModel(ctx);
-        List<TrafficJournalDao.TrafficData> results = null;
-
         String downloader = ctx.queryParam("downloader");
-        if (downloader == null || downloader.isBlank()) {
-            results = trafficJournalDao.getAllDownloadersOverallData(timeQueryModel.startAt(), timeQueryModel.endAt()).stream().peek(data -> fixTimezone(ctx, data)).toList();
-        } else {
-            results = trafficJournalDao.getSpecificDownloaderOverallData(downloader, timeQueryModel.startAt(), timeQueryModel.endAt()).stream().peek(data -> fixTimezone(ctx, data)).toList();
-        }
-
-        List<TrafficJournalDao.TrafficData> records = new ArrayList<>();
-        var it = results.iterator();
-        // -----
-        TrafficJournalDao.TrafficData base = null;
-        while (it.hasNext()) {
-            var target = it.next();
-
-            if (base == null) {
-                base = target;
-                continue; // 跳过插入当天的总数据，直接计算增量
-            }
-
-            long uploadedOffset = target.getDataOverallUploaded() - base.getDataOverallUploaded();
-            long downloadedOffset = target.getDataOverallDownloaded() - base.getDataOverallDownloaded();
-            if (uploadedOffset < 0) uploadedOffset = 0;
-            if (downloadedOffset < 0) downloadedOffset = 0;
-
-            // 插入增量数据
-            records.add(new TrafficJournalDao.TrafficData(target.getTimestamp(), uploadedOffset, downloadedOffset));
-            base = target;
-        }
-
-//        if (records.size() > 1) {
-//            // 多插的那个移除
-//            records.removeFirst();
-//        }
+        var records = trafficJournalDao.getDayOffsetData(downloader, timeQueryModel.startAt(), timeQueryModel.endAt(), d -> fixTimezone(ctx, d));
         ctx.json(new StdResp(true, null, records));
     }
